+++ src/minified-extras-full-src.js
@@ -418,12 +418,9 @@ function dummy() {
 		 * @requires set template
 		 * @configurable default
 		 * @name .ht()
-		 * @syntax list.ht(templateString)
-		 * @syntax list.ht(templateString, object)
-		 * @syntax list.ht(templateFunction)
-		 * @syntax list.ht(templateFunction, object)
-		 * @syntax list.ht(idSelector)
-		 * @syntax list.ht(idSelector, object)
+		 * @syntax list.ht(templateString, object...)
+		 * @syntax list.ht(templateFunction, object...)
+		 * @syntax list.ht(idSelector, object...)
 		 * @module WEB+UTIL
 		 * Replaces the content of the list elements with the HTML generated using the given template. The template uses
 		 * ##template() syntax and HTML-escaped its output using ##escapeHtml(). 
@@ -474,16 +471,18 @@ function dummy() {
 		 *                   of the specified &lt;script> element as template. This allows you to put your template into 
 		 *                   a &lt;script&gt; tag with a non-JavaScript type (see example). Any string that starts with '#' and does not
 		 *                   contain any spaces is used as selector.
-		 * @param object optional the object to pass to the template. If object is not set, the template is called with <var>undefined</var>
-		 *                        as object.
+		 * @param object optional one or more objects to pass to the template. If object is not set, the template is called with <var>undefined</var>
+		 *                        as object. If exactly one object is given, it is passed directly to the template. If you specify more than one 
+		 *                        object, they are ##merge#merged##.
 		 * @return the current list
 		 * 
 		 * @see ##HTML() creates only the nodes and can be used with ##add() and other methods to add the nodes to the DOM, giving you more flexibility than <var>ht()</var>.
 		 */
 		'ht': function(htmlTemplate, object) {
-			return this['set']('innerHTML', isFunction(htmlTemplate) ? htmlTemplate(object) : 
-				                            /{{/.test(htmlTemplate) ? formatHtml(htmlTemplate, object) : 
-				                            /^#\S+$/.test(htmlTemplate) ? formatHtml($$(htmlTemplate)['text'], object) : htmlTemplate);
+			var o = arguments.length > 2 ? merge(sub(arguments, 1)) : object;
+			return this['set']('innerHTML', isFunction(htmlTemplate) ? htmlTemplate(o) : 
+				                            /{{/.test(htmlTemplate) ? formatHtml(htmlTemplate, o) : 
+				                            /^#\S+$/.test(htmlTemplate) ? formatHtml($$(htmlTemplate)['text'], o) : htmlTemplate);
 		 }
 		/*$
 		 * @stop
@@ -662,12 +661,9 @@ function dummy() {
 		 * @requires template ht
 		 * @configurable default
 		 * @name HTML()
-		 * @syntax HTML(templateString)
-		 * @syntax HTML(templateString, object)
-		 * @syntax HTML(templateFunction)
-		 * @syntax HTML(templateFunction, object)
-		 * @syntax HTML(idSelector)
-		 * @syntax HTML(idSelector, object)
+		 * @syntax HTML(templateString, object...)
+		 * @syntax HTML(templateFunction, object...)
+		 * @syntax HTML(idSelector, object...)
 		 * @module WEB
 		 * Creates a ##list#list## of HTML nodes from the given HTML template. The list is compatible with ##add(), ##fill() and related methods.
 		 * The template uses the ##template() syntax with ##escapeHtml() escaping for values.
@@ -721,14 +717,17 @@ function dummy() {
 		 *                   of the specified &lt;script> element as template. This allows you to put your template into 
 		 *                   a &lt;script&gt; tag with a non-JavaScript type (see example). Any string that starts with '#' and does not
 		 *                   contain any spaces is used as selector.
-		 * @param object optional the object to pass to the template
+		 * @param object optional one or more objects to pass to the template. If object is not set, the template is called with <var>undefined</var>
+		 *                        as object. If exactly one object is given, it is passed directly to the template. If you specify more than one 
+		 *                        object, they are ##merge#merged##.
 		 * @return the list containing the new HTML nodes
 		 *  
 		 * @see ##ht() is a shortcut for <code>fill(HTML())</code>.
 		 * @see ##EE() is a different way of creating HTML nodes.
 		 */
-		'HTML': function (htmlTemplate, object) {
-		    return  _(EE('div')['ht'](htmlTemplate, object)[0].childNodes);
+		'HTML': function () {
+			var div = EE('div');
+		    return  _(call(div['ht'], div, arguments)[0].childNodes);
 		},
 		/*$
 		 * @stop